+++ src/lib/features/cart/cartSlice.ts
@@ -22,32 +22,28 @@ const cartSlice = createSlice({
   name: 'cart',
   initialState,
   reducers: {
-    // Add item to the cart (increase quantity if item exists)
     addToCart: (state, action: PayloadAction<Omit<CartItem, 'quantity'>>) => {
       const existingItem = state.items.find(
         item => item.id === action.payload.id
       );
 
       if (existingItem) {
-        existingItem.quantity += 1; // Increase quantity if item already in cart
+        existingItem.quantity += 1;
       } else {
-        state.items.push({ ...action.payload, quantity: 1 }); // Add new item with quantity 1
+        state.items.push({ ...action.payload, quantity: 1 });
       }
 
-      state.isCartOpen = true; // Open cart when adding item
+      state.isCartOpen = true;
     },
 
-    // Remove item from the cart by ID
     removeFromCart: (state, action: PayloadAction<string>) => {
       state.items = state.items.filter(item => item.id !== action.payload);
     },
 
-    // Clear all items in the cart
     clearCart: state => {
       state.items = [];
     },
 
-    // Update quantity of an item in the cart
     updateQuantity: (
       state,
       action: PayloadAction<{ id: string; quantity: number }>
@@ -55,11 +51,10 @@ const cartSlice = createSlice({
       const item = state.items.find(item => item.id === action.payload.id);
 
       if (item && action.payload.quantity > 0) {
-        item.quantity = action.payload.quantity; // Set new quantity if valid
+        item.quantity = action.payload.quantity;
       }
     },
 
-    // Toggle the visibility of the cart
     toggleCart: (state, action: PayloadAction<boolean>) => {
       state.isCartOpen = action.payload;
     }
